+++ controllers/auth-controllers.js
@@ -1,8 +1,7 @@
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 const gravatar = require("gravatar");
-
-
+const fs = require('fs/promises');
 require("dotenv").config();
 const { User } = require("../models/user");
 const { HttpError } = require("../helpers");
@@ -10,7 +9,22 @@ const { ctrlWrapper } = require("../decorators");
 
 const { SECRET_KEY } = process.env;
 
+const path = require("path");
 
+const avatarsDir = path.resolve("public", "avatars");
+
+const updateAvatar = async (req, res) => {
+  const { _id } = req.user;
+  const { path: oldPath, filename } = req.file;
+  const newPath = path.join(avatarsDir, filename);
+  await fs.rename(oldPath, newPath);
+  const avatarURL = path.join("avatars", filename);
+  await User.findByIdAndUpdate(_id, { avatarURL });
+
+  res.json({
+    avatarURL
+  });
+};
 
 const register = async (req, res) => {
   const { email, password } = req.body;
@@ -75,5 +89,5 @@ module.exports = {
   login: ctrlWrapper(login),
   getCurrent: ctrlWrapper(getCurrent),
   logout: ctrlWrapper(logout),
- 
+  updateAvatar: ctrlWrapper(updateAvatar),
 };
