+++ controllers/contacts-controllers.js
@@ -1,26 +1,11 @@
-const fs = require('fs').promises;
+// const fs = require('fs').promises;
 const { HttpError } = require("../helpers");
 const { Contact } = require("../models/contact");
 const { ctrlWrapper } = require("../decorators");
 
-const { User } = require("../models/user");
+// const { User } = require("../models/user");
 
-const path = require("path");
 
-const avatarsDir = path.join(__dirname, "../", "public", "avatars");
-
-const updateAvatar = async (req, res) => {
-  const { _id } = req.user;
-  const { path: tempUpload, filename } = req.file;
-  const resultUpload = path.join(avatarsDir, filename);
-  await fs.rename(tempUpload, resultUpload);
-  const avatarURL = path.join("avatars", filename);
-  await User.findByIdAndUpdate(_id, { avatarURL });
-
-  res.json({
-    avatarURL
-  });
-};
 
 const getAllContacts = async (req, res) => {
   const { _id: owner } = req.user;
@@ -71,5 +56,5 @@ module.exports = {
   addContact: ctrlWrapper(addContact),
   removeContact: ctrlWrapper(removeContact),
   updateContactById: ctrlWrapper(updateContactById),
-  updateAvatar: ctrlWrapper(updateAvatar)
+  // updateAvatar: ctrlWrapper(updateAvatar)
 };
