+++ models/contact.js
@@ -30,7 +30,7 @@ const contactsSchema = new Schema(
     },
     owner: {
       type: Schema.Types.ObjectId,
-      ref: "user",  
+      ref: "user"
     }
   },
   { versionKey: false }
